+++ tests/run.clj
@@ -76,6 +76,7 @@
 (require "tests/data_readers_test")
 (require "tests/spec_test")
 (require "tests/host_threads_foundation_test")
+(require "tests/capability_metadata_test")
 (require "tests/async_buffer_test")
 (require "tests/async_api_test")
 (require "tests/async_alts_test")